Refactor: Break up client main function
