+++ client/src/lib.rs
@@ -1,10 +1,10 @@
+pub mod eventloop;
 pub mod util;
 pub mod input;
 pub mod loaders;
 pub mod networking;
 pub mod renderer;
 pub mod simulation;
-pub mod window;
 
 mod client;
