+++ client/src/renderer/mod.rs
@@ -1,13 +1,18 @@
+pub mod config;
 pub mod mesh;
 pub mod model;
+pub mod renderer;
 pub mod scene;
 pub mod terrain;
 pub mod texture;
+pub mod window;
 
 mod graph;
 mod pipeline;
 
+pub use config::DisplayConfig;
 pub use graph::RenderGraph;
 pub use mesh::Mesh;
 pub use model::Model;
+pub use renderer::Renderer;
 pub use texture::Texture;
